+++ iree/turbine/kernel/boo/__init__.py
@@ -6,4 +6,3 @@
 
 from . import modeling
 from . import ops
-from . import conv_exports
