--- iree/turbine/kernel/boo/conv_exports/__init__.py
@@ -1,11 +1,0 @@
-"""Tools for generating forward and backward convolution kernels."""
-
-# Copyright 2025 The IREE Authors
-#
-# Licensed under the Apache License v2.0 with LLVM Exceptions.
-# See https://llvm.org/LICENSE.txt for license information.
-# SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
-
-from .conv import *
-from .utils import *
-from .preload import *
